Add Gum installation to setup wizard
This commit adds a new section to the setup wizard that asks the user for permission to download and run Gum. Gum is a terminal UI that provides a better experience for command-line applications. The script will exit if the user denies permission, as Gum is required to run the script.